Formations superclass created with subclass X_formation and Diamond_formation
